Create covid_data_sql_exploration
This project shows some of my aptitude using SQL. 
The data used in this is accurate up to 21st December, 2021. 
Unfortunately, there is no proper breakdown with each variant of covid19 so we can examine further.
More could have been done as there is so much information to sift through and find meaningful insights from, but for the purposes of demonstration this is sufficient.
Thank you for reading. 
Comments and suggestions are welcome.